moving dynamic label function to helper
