+++ gitApp/hook/helper.py
@@ -1,4 +1,6 @@
 import json
+from models.Results import Results
+
 
 VERBOSE = True
 
@@ -134,3 +136,15 @@ def convertOutput(out):
         s = f'ERROR CONVERTING SYS OUT:\n{e}\n{out}'
         print('output conversion failed', e, out)
     return s
+
+
+def get_dyn_keys(res: Results):
+    out = ''
+    all_attr = res.__dict__
+    keys = all_attr['_fields_ordered']
+    for k in keys:
+        if 'dynamic_' in k:
+            out += k.replace('dynamic_', '') + ": "
+            out += str(all_attr[k]) + ' '
+    out.rstrip(' ')
+    return out
